useBoardColumns db call addition
